+++ CHANGELOG.md
@@ -125,6 +125,7 @@
 - switch from /bin/bash to /usr/bin/env to allow more setups like nixos !2708
 - Use one database dump before tests so that the tests are not depending on each other !2701 @koemai
 - Fix deprecated strtolower in MailsControl.php #1569 !2718 @stcz
+- Added security page !2764 @martincodes_de
 
 ## Development documentation